% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{breastcancerdiagnostic}
\alias{breastcancerdiagnostic}
\title{Breast cancer diagnostic data}
\format{
A data frame with 569 rows and 31 variables:
\describe{
\item{radius_mean}{Mean radius.}
\item{texture_mean}{Mean texture.}
\item{perimeter_mean}{Mean perimeter.}
\item{area_mean}{Mean area.}
\item{smoothness_mean}{Mean smoothness.}
\item{compactness_mean}{Mean compactness.}
\item{concavity_mean}{Mean concavity.}
\item{concave_pts_mean}{Mean number of concave points.}
\item{symmetry_mean}{Mean symmetry.}
\item{fractal_dim_mean}{Mean fractal dimension.}
\item{radius_se}{Radius standard error.}
\item{texture_se}{Texture standard error.}
\item{perimeter_se}{Perimeter standard error.}
\item{area_se}{Area standard error.}
\item{smoothness_se}{Smoothness standard error.}
\item{compactness_se}{Compactness standard error.}
\item{concavity_se}{Concavity standard error.}
\item{concave_pts_se}{Concave points standard error.}
\item{symmetry_se}{Symmetry standard error.}
\item{fractal_dim_se}{Fractal dimension standard error.}
\item{radius_worst}{Worst radius.}
\item{texture_worst}{Worst texture.}
\item{perimeter_worst}{Worst perimeter.}
\item{area_worst}{Worst area.}
\item{smoothness_worst}{Worst smoothness.}
\item{compactness_worst}{Worst compactness.}
\item{concavity_worst}{Worst concavity.}
\item{concave_pts_worst}{Worst number of concave points.}
\item{symmetry_worst}{Worst symmetry.}
\item{fractal_dim_worst}{Worst fractal dimension.}
\item{diagnosis}{Diagnosis outcome (\code{"B"} = benign, \code{"M"} = malignant).}
}
}
\source{
Breast Cancer Wisconsin Diagnostic Dataset from the UCI Machine
Learning Repository, packaged in \code{dslabs::brca}.
}
\usage{
breastcancerdiagnostic
}
\description{
A binary classification dataset for breast cancer diagnosis using tumor
morphology measurements.
}
\details{
Column names were standardized to \code{snake_case} when packaging the data.
}
\examples{
str(funcml::breastcancerdiagnostic)
table(funcml::breastcancerdiagnostic$diagnosis)
}
\keyword{datasets}
