% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{breastcancerwisconsin}
\alias{breastcancerwisconsin}
\title{Wisconsin breast cancer data}
\format{
A data frame with 699 rows and 10 variables:
\describe{
\item{clump_thickness}{Clump thickness score.}
\item{uniformity_cell_size}{Uniformity of cell size score.}
\item{uniformity_cell_shape}{Uniformity of cell shape score.}
\item{marginal_adhesion}{Marginal adhesion score.}
\item{single_epithelial_cell_size}{Single epithelial cell size score.}
\item{bare_nuclei}{Bare nuclei score.}
\item{bland_chromatin}{Bland chromatin score.}
\item{normal_nucleoli}{Normal nucleoli score.}
\item{mitoses}{Mitoses score.}
\item{class}{Diagnostic class (\code{2} = benign, \code{4} = malignant).}
}
}
\source{
Wisconsin Breast Cancer Database from University of Wisconsin
Hospitals, distributed through the UCI Machine Learning Repository.
}
\usage{
breastcancerwisconsin
}
\description{
A binary classification dataset for breast cancer diagnosis from cytology
measurements.
}
\examples{
str(funcml::breastcancerwisconsin)
table(funcml::breastcancerwisconsin$class)
}
\keyword{datasets}
