% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{group_cv}
\alias{group_cv}
\title{Grouped cross-validation.}
\usage{
group_cv(v = 5, group, repeats = 1, seed = NULL)
}
\arguments{
\item{v}{Number of folds.}

\item{group}{Grouping variable name or vector.}

\item{repeats}{Number of repeats.}

\item{seed}{Optional seed.}
}
\value{
A \code{funcml_cv} object.
}
\description{
Grouped cross-validation.
}
