% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{haberman}
\alias{haberman}
\title{Haberman survival data}
\format{
A data frame with 306 rows and 4 variables:
\describe{
\item{age}{Age of patient at operation time in years.}
\item{operation_year}{Year of operation minus 1900.}
\item{positive_axillary_nodes}{Number of positive axillary nodes detected.}
\item{survival_status}{Survival status (\code{1} = survived 5 years or longer,
\code{2} = died within 5 years).}
}
}
\source{
Haberman's Survival Data from the University of Chicago's Billings
Hospital study, distributed through the UCI Machine Learning Repository.
}
\usage{
haberman
}
\description{
A binary classification dataset on breast cancer survival after surgery.
}
\examples{
str(funcml::haberman)
table(funcml::haberman$survival_status)
}
\keyword{datasets}
