% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{holdout}
\alias{holdout}
\title{Plain holdout resampling.}
\usage{
holdout(prop = 0.8, strata = TRUE, seed = NULL)
}
\arguments{
\item{prop}{Training-set proportion.}

\item{strata}{Logical; stratify classification outcomes.}

\item{seed}{Optional seed.}
}
\value{
A \code{funcml_cv} object.
}
\description{
Plain holdout resampling.
}
\examples{
holdout(prop = 0.75, seed = 1)
}
