% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{list_metrics}
\alias{list_metrics}
\title{List available metrics used in scoring and resampling summaries.}
\usage{
list_metrics(direction = NULL, columns = NULL)
}
\arguments{
\item{direction}{Optional character filter: \code{"maximize"} or \code{"minimize"}.}

\item{columns}{Optional character vector of columns to return.}
}
\value{
Data frame of metric metadata.
}
\description{
List available metrics used in scoring and resampling summaries.
}
\examples{
list_metrics()
list_metrics(direction = "minimize")
}
