% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{list_tunable_learners}
\alias{list_tunable_learners}
\title{Shortcut for learners with tuning support.}
\usage{
list_tunable_learners(...)
}
\arguments{
\item{...}{Passed to \code{\link[=list_learners]{list_learners()}}.}
}
\value{
Data frame with the same columns as \code{\link[=list_learners]{list_learners()}}.
}
\description{
Shortcut for learners with tuning support.
}
\examples{
list_tunable_learners()
}
