% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{mammography}
\alias{mammography}
\title{Mammography calcification data}
\format{
A data frame with 11,183 rows and 7 variables:
\describe{
\item{attr1}{Numeric imaging-derived predictor 1.}
\item{attr2}{Numeric imaging-derived predictor 2.}
\item{attr3}{Numeric imaging-derived predictor 3.}
\item{attr4}{Numeric imaging-derived predictor 4.}
\item{attr5}{Numeric imaging-derived predictor 5.}
\item{attr6}{Numeric imaging-derived predictor 6.}
\item{class}{Calcification class (\verb{\\"-1\\"} or \verb{\\"1\\"}).}
}
}
\source{
Woods K, Doss C, Bowyer K, Solka J, Priebe C, Kegelmeyer P
(1993). Comparative evaluation of pattern recognition techniques for
detection of microcalcifications in mammography.
}
\usage{
mammography
}
\description{
A binary classification dataset for detection of mammographic
microcalcifications.
}
\examples{
str(funcml::mammography)
table(funcml::mammography$class)
}
\keyword{datasets}
