% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{pimadiabetes}
\alias{pimadiabetes}
\title{Pima diabetes data}
\format{
A data frame with 532 rows and 9 variables:
\describe{
\item{npreg}{Number of pregnancies.}
\item{glu}{Plasma glucose concentration.}
\item{bp}{Diastolic blood pressure.}
\item{skin}{Triceps skin fold thickness.}
\item{bmi}{Body mass index.}
\item{ped}{Diabetes pedigree function.}
\item{age}{Age in years.}
\item{diabetes}{Diabetes outcome (\code{"Yes"} or \code{"No"}).}
\item{split}{Suggested split indicator (\code{"train"} or \code{"test"}).}
}
}
\source{
National Institute of Diabetes and Digestive and Kidney Diseases
Pima Indians Diabetes Database.
}
\usage{
pimadiabetes
}
\description{
A diabetes classification dataset with clinical measurements and a predefined
train/test split column.
}
\examples{
str(funcml::pimadiabetes)
table(funcml::pimadiabetes$split)
}
\references{
Smith JW, Everhart JE, Dickson WC, Knowler WC, Johannes RS
(1988). Using the ADAP learning algorithm to forecast the onset of
diabetes mellitus. In \emph{Proceedings of the Annual Symposium on Computer
Application in Medical Care}, 261-265.
}
\keyword{datasets}
