% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_proc.R
\name{roc-methods}
\alias{roc-methods}
\alias{print.funcml_roc}
\alias{plot.funcml_roc}
\title{Methods for ROC curve results.}
\usage{
\method{print}{funcml_roc}(x, digits = 4L, ...)

\method{plot}{funcml_roc}(x, ...)
}
\arguments{
\item{x}{A \code{funcml_roc} object.}

\item{digits}{Number of digits numeric columns are rounded to when printed.}

\item{...}{Additional arguments (unused).}
}
\value{
\code{print()} returns the input object invisibly. \code{plot()} returns a
\code{ggplot2} object.
}
\description{
Methods for ROC curve results.
}
