% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{time_cv}
\alias{time_cv}
\title{Time-aware rolling resampling.}
\usage{
time_cv(
  initial,
  assess = 1,
  time = NULL,
  skip = 0,
  cumulative = TRUE,
  seed = NULL
)
}
\arguments{
\item{initial}{Initial training window size.}

\item{assess}{Assessment window size.}

\item{time}{Ordering variable name or vector.}

\item{skip}{Number of observations to skip between splits.}

\item{cumulative}{Logical; use an expanding training window.}

\item{seed}{Optional seed.}
}
\value{
A \code{funcml_cv} object.
}
\description{
Time-aware rolling resampling.
}
