% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune}
\alias{tune}
\title{Simple grid search tuning.}
\usage{
tune(
  data,
  formula,
  model,
  grid,
  resampling = cv(5),
  metric = NULL,
  type = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame.}

\item{formula}{Model formula.}

\item{model}{Learner id.}

\item{grid}{Data frame of hyperparameter combinations.}

\item{resampling}{Resampling object.}

\item{metric}{Metric to optimize.}

\item{type}{Prediction type override.}

\item{seed}{Optional seed.}

\item{...}{Passed to \code{fit()}.}
}
\value{
A \code{funcml_tune} object.
}
\description{
Simple grid search tuning.
}
